+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 * Show relevant step by step nav based on user journey (PR #501)
 * Add Error alert component (PR #503)
+* Add experimental panel component based on GOV.UK Frontend (PR #507)
 
 ## 9.17.1